Update thesaurus.py, Update examples.py, Update README.md
Moved the code that was in getWord to findWordTotal, as the code that *was* in findWordTotal didn't work, and I thought getWord was somewhat of an obscure name. They both did the same thing, but the original getWord 1) actually worked and 2) had better looking code.

I also added a tutorial on its general use in the README, and added Kyle to the special thanks. The example I gave him is also in examples.py now.

I think the next TODO to work on will be making this stuff into an easy to use class.