Create Word class, Add filtering, Update README.md
Made it so you can do your nice filtering now. Next thing to do will be to add more comments to the source and also probably fix the documentation to show that when you look at a single definition with .synonyms(), you only get a list of the synonyms. You don't get a list of a list.

As always, enjoy :)